Added day 35
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible
Mesage posted in the Ukraine IT Army for Day 35:
Good morning! 
Let’s shift the focus to the job listing websites. We are going to have quite a few targets and they are not that simple. So let’s put all our efforts into action!