Added day 36
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible
Mesage posted in the Ukraine IT Army for Day 36:
IT Army, our target of the day is government information systems for the veterinary industry “VetIS”. They are used by everyone who either produces or sells products that are subject to veterinary control. It’d be a significant achievement if we manage to knock them off. So let’s get down to business!