Added day 37
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible
Mesage posted in the Ukraine IT Army for Day 37:
Good morning, we are from Ukraine!
Administrators of the targeted platforms are actively blocking and filtering our traffic. In the evening they even managed to switch back from the emergency mode. 
Today we won’t switch to new targets and will continue working on yesterday's ones. We’d ask for your full involvement in the attacks. Moreover, we need to use only Russian VPN addresses and their local proxy! They closed the system for external access but we need to continue the attack. Just a reminder that our targets are the systems where all Russian animal origin food producers, wholesale and trading companies manage certification and ECM tracking.