Added Day 45
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 45:
During the weekend, we kept the online TV platforms down. It's time to switch back to the services that are vital for business. Today we'll work with the following targets: