Added Day 46
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 46:
Yesterday was the last day of the aggressor country's Ministry of Trade resolution not to mark certain goods. It's high time to remind them about us.
We are launching a new wave of attacks on their marking system.
You know that we've done it before, so it's mandatory to use a VPN (ideally, the aggressor country's one).