Added Day 77 + EGAIS
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 77:
Obviously, import substitution in tech doesn't work well in Russia. Rutube (a local imitation of Youtube) followed the lead of the Russian warship. So let's send other Russian replications of Western services in the same direction! Below is the first batch of targets.
Just a quick reminder that attacks guidelines and other information can be found on our website: https://itarmy.com.ua/instruction/?lang=en
EGAIS:
"EGAIS" is a Russian federal state-owned automatic system for tracking the production and distribution of alcoholic beverages in the country. If this system is down, the official turnover of liquor and spirits in Russia will be blocked!