Fixed Day 78 and Added Day 79
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 79:
Today let's help an allied community with their attack on resources related to aviation and airlines. This is just a reminder that Russia stole airplanes that were on lease at a total cost of 20 billiard dollars.