Added Day 82
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 82:
The working week for Russians should start with malfunctioning digital signatures. Below is a list of OCSP resources that provide those services in Russia.