Added Day 89
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 89:
You did a great job over the weekend! Based on the comments on social media, online banking and mobile apps have been down for three days now. To increase our coverage, let's also target microfinance institutions.
Below is a link to the updated list of targets. If you use "Death by a 1000 needles" or mhddos_proxy, their updated configurations will include the most up-to-date targets out-of-the-box. And don't forget to tag --itarmy when launching the mhddos_proxy.