Added Day 91
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 91:
Let's move on with the Russian financial system. Let's switch to stock exchanges and analytical portals.
Banks and microfinance institutions under attack remain a valid target.
And please upgrade mhddos_proxy to the latest version!