Added Day 105
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 105:
We've added more russian government web resources to our list. Thanks a lot to everyone who searches and shares with us future targets via chat or email!