Added Day 110
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 110:
No list was posted today, recycling yesterdays targets:
We continue to attack the russian media, government web resources, online platforms. We also remind the occupiers that Crimea is Ukraine!