Added Day 120
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 120:
Public services the underbelly of digital Russia to be under the aim today again. This one is huge touching hundreds conjunctive services interfered in economy and military. All other targets be paused to be sure it is taken down.