Added Day 123
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 123:
We are onto "corrective labor" for russian media of the cross-border areas today. Spoiler alert: how to make negative economy growth in russia is the next week topic.