Added Day 134
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 134:
We finally managed to put together logistics. The targets cover both civil and military supply. We are going to slooooow them down today!