Added Day 135
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 135:
Part 2 of logistics. The targets cover both civil and military supply again, We are going to slooooow them down again.