Added Day 143
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 143:
No new targets posted for this day, recycling yesterdays targets. Beeline, the telecom provider, is the top-notch target for you today. We made a few testing pinches, and Beeline responded swiftly :) Please keep rolling on Foreign Affairs services.