Added Day 146
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 146:
We continue to prevent equip the orcs with drones. Your recommendations have been added to the list of targets.
You can grant access to your cloud resources to our bot that will launch a coordinated attack from all the available servers. It has been improved and it is already possible to transfer servers with a non-standard SSH port.