Added Day 158
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 158:
No new message or targets posted today, recycling previous targets. Today we'll start working on russian online banking. We also ask you to share information about the resources with which residents of the russian federation bypass sanctions.