Added Day 166
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 166:
We challenge import phaseout players of the aggressor. TrueConf is the platform for video calls and remote works. It is quite popular among private and public companies in russia recently.