Added Day 179
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 179
No new list today (yet), recycling previous targets with new set of Targets in telegram by the IT Army of Ukraine, will update if new IP list is released:
The work with the central bank of the russian federation was so successful that they had to disable some services! Moreover, the more protected networks of the central bank could not cope with our attack, so the russians had to completely transfer everything to the service provider.
As a result, import substitution went "well" as usual and confirmed once again that the products of the russian federation are at such a low level that they have no analogues.
Today we have a rehearsal with banking services that still allow russians to issue international virtual cards and send money.