Added Day 187
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 187
No new list today (yet), recycling previous targets with new set of Targets in telegram by the IT Army of Ukraine, will update if new IP list is released:
You did it great and we tend to take a look on another popular money transfer services koronapay.com and yoomoney.ru. The sanctions implied on Russia are good but often worked around. Should we help them out?