Added Day 211
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
no new Message posted in the Ukraine IT Army for Day 211, recycling previous targets:
The escalation of the war by putin has a negative effect on the quotations on the moscow stock exchange every time. Let's make it difficult for them by attacking their network early in the trade.