Added Day 212
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New Message posted in the Ukraine IT Army for Day 212:
Now even the countries, which according to the rashists are "friendly", stop accepting Mir bank cards. Today we will attack the infrastructure networks of Mir/NSPK. Our goal is to ensure that in russia there are also interruptions in the processing of payments and in the operation of terminals.