Added Day 218
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
new Message posted in the Ukraine IT Army for Day 218, however target URL's not updated, recycling previous targets:
Today, we aim to create problems in making payments and transfers in the sanctioned Sovkombank. We have already attacked it, and the second time it may be more difficult and resource-intensive, so we invite everyone to join!