Added Day 222 Part 2
New Targets for 2nd half of Day 222
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 222 Part 2:
The results are there, but to completely shutdown them on the servers, changes to the plan are needed. The target is undergoing correction. Now we are switching to another chain of online stores - mvideo