Added Day 223
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 223:
You did a good job with mvideo yesterday. After the necessary adjustments, we return to citilink. Join the attack!