Added Day 225
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 225:
Cartographic and geographic information systems are the most important infrastructural element of warfare and logistics. One of the largest such systems of the aggressor country is 2GIS. Since the beginning of the war, they have been informing the aggressor state about the current routes, the situation on the roads, crowds, obstacles and roadblocks. The company belongs to a sub-sanctioned Sberbank, the application was removed from the AppStore