Added Day 233 (normal | https only | http only)
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 233:
Today we return to banks again. Tinkoff Bank in the sights.
We aim to interrupt payment processing, delay liabilities, undermine bank’s reputation as usual.