Added Day 248 and updated List for Days (bulk TL)
Update for days and list of Days, List for Days Updated (TL)
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Targets recycled - New message was posted in the Ukraine IT Army for Day 248:
Customers of the Russian Alfa-Bank cannot make payments, or at least log in to online banking or the application. Also they can't get into Alfa-investment either. We notice more and more comments about problems, so today we continue to work on the Alfa infrastructure.
Alfa says that "there is no failure". They remind someone... 🤔