Updated Day 277 - Targets (hidden) & days (TL)
Updated targets for days, simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new message posted in the Ukraine IT Army for Day 277, recycling previous day:
The lack of electricity and the Internet does not stop our work. We are ready for it. We end the working week with a repeated attack on the russian Alfabank. Let their customers sit without shopping on Black Friday.