Added Day 415 - Targets and Days (TL) Eng+Ukr
Updated targets for days, simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible.
Message Posted in the IT Army Telegram server for current targets:
The day started badly for Tinko, which supplies security systems and RESO insurance. The first has recovered a little from our blow, while the insurer remains firmly in place. At the same time, we are working on an important goal, which we will tell you about later. Turn everything on, every megabit counts now!
\**\ 
The Russian Alfastrakh insurance company is not working, claiming that the internet has run out. It's the karmic fate of the aggressor's insurance business. It's a pity that there's no electricity, but considering the direction of their country's movement, it's only a matter of time. Well done, ITtackers, great shot!
/ *** / 
День почався недобре для Тінко, що постачає охоронні системи, та РЕСО-страхування. Перший трошки відійшов від нашого удару, а страховик лежить міцно. Паралельно працюємо над важливою ціллю, про яку розповімо трошки пізніше. Вмикайте все, зараз кожен мегабіт в допомогу!