Added Day 433 - Targets and Days (TL) Eng+Ukr
Updated targets for days, simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible.
Message Posted in the IT Army Telegram server for current targets:
Today, clients of Aktіon (glavbukh.ru and buhsoft.ru) couldn't find their balance. Our efforts to negatively fill the aggressor's budget with taxes continue. The main thing is to see how they immediately switched to the internet provider. Lying is their thing.
Oh, here is that juicy thing we've been helping and still helping to lay down - the enemy's customs. We managed to completely assemble such a machine only once, but individual services suffer regularly. And we haven't finished yet ;)
In parallel, we continue to work on accounting - today nalog-nalog.ru and assistentus.ru went for "reconciliation."
/ *** / 
Сьогодні вже не знайшли сальдо клієнти Актіон (glavbukh.ru та buhsoft.ru). Наші зусилля по від'ємному наповненню податками бюджета агресора продовжуються. Ви головне подивіться, як вони одразу з'їхали на інтернет-провайдера. Брехати таки їхнє все.
О ось і те жирненьке, що ми допомагали і досі допомагаємо покласти - митниця ворога. Повністю скласти таку машину вдалося тільки раз, але окремі сервіси страждають регулярно. І ми ще не закінчили ;)
Паралельно продовжуємо працювати по бухгалтерському обліку - сьогодні nalog-nalog.ru та assistentus.ru пішли на "переоблік".