Added Day 522 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
"Political divergences and wars are opening up new fields for cyber attacks. A case in point is the IT Army of Ukraine, which has launched targeted operations against Russian websites, including attacks on RZD. The significance of geopolitical events in cyberspace is hard to overstate. The integration of cybersecurity with geopolitics is a bridge between the real world and the digital one." writes British TEISS.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Повідомлення оприлюднило IT Army of Ukraine:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub. Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут. Повідомлення оприлюднило IT Army of Ukraine:
"Політичні розбіжності та війни відкривають нові поля для кібератак. Прикладом є IT Army of Ukraine, яка розпочала цілеспрямовані операції проти російських сайтів, включаючи атаки на РЖД. Важливість геополітичних подій у кіберпросторі важко переоцінити. Інтеграція кібербезпеки з геополітикою - це міст між реальним світом та цифровим." пише британська TEISS.