Updated info box to include links for day 524 | Оновлено інформаційне поле, щоб включити посилання на день 524
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Comrade Lyapunov, the very VP of 'Rostec', calls to believe in miracles: 'In the past year and a half, there hasn't been a single successful attack on Russia's 'Gosuslugi’.” Do you recall our 'unsuccessful' attacks back in March '22? And what about the recent hit on regional 'Gosuslugi' websites on June 23rd, 2023? Everything was dead in the water. Maybe Lyapunov should pay our channel more frequent visits to stay in the loop?
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Повідомлення оприлюднило IT Army of Ukraine:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub. Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут. Повідомлення оприлюднило IT Army of Ukraine:
Товариш Ляпунов, той що віцепрезидент "Роскосмоса", закликає вірити в чудеса: "За півтора роки не було жодної успішної атаки на Госуслуги расії”. Ви пам'ятаєте наші 'неуспішні' атаки ще в березні '22? А як щодо нещодавньої атаки на регіональні сайти держпослуг 23 червня 2023? Лежало все намертво. Можливо Ляпунову варто частіше відвідувати наш канал, щоб бути в курсі подій?