Updated InfoBoxes & added links for day 539 | Оновлено інформаційні блоки та додано посилання для дня 539
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Lugakom is still reeling, and we're already hitting "PEK," a respected freight carrier from the Ruzzia. Logistics complications often affect not only the economy but also the military, who transport components for weapon manufacturing through such companies. That's why such targets are set in the mission of the IT ARMY. Have you already joined our missions with the new software "UkITA," which even your grandma can use?
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Щоб не відставати від цілей, ми збираємося одночасно використовувати цілі db1000n із їхнього сховища GitHub. Ці оновлення відбуваються щодня, тож якщо db1000n зміниться, ймовірно, пройде ще кілька років, перш ніж зміни відобразяться тут. Повідомлення оприлюднило ІТ-Армія України:
Лугаком досі не оговтався, а ми вже б'ємо по «ПЭК», поважному перевізнику вантажів на роззії. Ускладнення логістики часто впливає не тільки на економіку, але і на військових, які перевозять компоненти для виробництва зброї саме такими компаніями. Тому такі цілі й закріплені у місії IT ARMY. А ви вже приєднались до наших місій з новим ПО "UkITA", яке зможе використовувати навіть ваша бабуся?