Added download PDF button to day 552 | До дня 552 додано кнопку завантаження PDF
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Do you remember our attack on Dom.ru provider? After that, our hackers discovered a vulnerability in the provider's network security. The scale of the damage is such that they came to make a complete replacement of routers, reinstall systems and restore databases. It also significantly affected the work of their clients, including the FSB.
In addition, we publish documents that show how the FSB controls the digital infrastructure and has access to a lot of data about the residents of the aggressor country.
Join IT ARMY chat.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Пам'ятаєте нашу атаку на провайдера Дом.ру? Поза цим, наші хакери знайшли вразливості в мережевому обладнанні провайдера. Масштаби збитків такі, що прийшлося робити повну заміну маршрутизаторів, перевстановлювати системи та відновлювати бази даних. Це також суттєво вплинуло на роботи їх клієнтів, серед яких і ФСБ.
Ще публікуємо документи які показують як ФСБ контролює цифрову інфраструктуру, і має доступ до безлічі даних про мешканців країни-агресора.
Приєднуйтесь до чату IT ARMY.