Added Day 556 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Do you remember our attack on Dom.ru provider? After that, our hackers discovered a vulnerability in the provider's network security. The scale of the damage is such that they came to make a complete replacement of routers, reinstall systems and restore databases. It also significantly affected the work of their clients, including the FSB.
In addition, we publish documents that show how the FSB controls the digital infrastructure and has access to a lot of data about the residents of the aggressor country.
Join IT ARMY chat.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Платіжні системи Фактура та Корона стали не здатні обслуговувати клієнтів зі вчора. Напевно таки щури проскочили. Нічого, може ворог ще до ідей комунізму повернеться, а там і гроші стануть не потрібні.
Ви знали що ви можете встановити ADSS на свій Raspberry і приєднатися до наших місій?