Updated Info box for day 558 | Оновлено інформаційне поле для дня 558
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
This time, 'The British Army Review,' a magazine published by The Centre for Historical Analysis and Conflict Research, has highlighted the unique approach of IT ARMY in participating in modern conflict.
We're not just volunteers; we're innovators changing the rules of the game. Our unity, strategy, and numbers make us a force to be reckoned with. Want to know how the world sees us? Dive into the article. Remember, this isn't just a story about us; it's a testament to what we stand for.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Цього разу "The British Army Review", журнал який випускає The Centre for Historical Analysis and Conflict Research, виділило унікальний підхід IT ARMY до участі в сучасному конфлікту. 
Ми не просто добровольці; ми новатори, які змінюють правила гри. Наша єдність, стратегія та чисельність роблять нас силою, з якою треба рахуватися. Хочете дізнатися, як нас бачить світ? Занурюйтесь у статтю. Пам'ятайте, це не просто історія про нас; це свідчення того, за що ми стоїмо.