Updated info box for Day 565 | Оновлено інформаційне поле для дня 565
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
When the Ruzzian MFA points fingers at the IT Army for being backed by global 'Sofa Hackers,' we can't help but smile. They've finally realized what we've known all along: we're not just a local phenomenon, we're a global force. The term 'Sofa Hacker' isn't an insult; it's a badge of honor that even reputable institutions like the Center for Strategic and International Studies are adopting. So to our community around the world, keep doing what you're doing. You've got their attention; now let's keep the momentum.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Доки МЗС розії вказує пальцем на IT Армію, називаючи нас всесвітніми 'диванними хакерами', ми не можемо не посміхнутися. Вони нарешті зрозуміли те, що ми вже давно знали: ми не просто локальне явище, ми глобальна сила. Термін 'диванний хакер' – це не образа, це звання честі, яке навіть такі авторитетні організації як Центр стратегічних та міжнародних досліджень вже використовують. Тому нашій глобальній спільноті: продовжуйте робити те, що ви робите. Ви привернули їх увагу; тепер не зупиняймося.