Updated Info box for day 268 | Оновлена інформаційна скринька на день 268
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
"It's all the more important to focus on Kyiv's main 'strike force' — the so-called Ukrainian 'IT Army.' Formed supposedly on a volunteer basis, this structure is a serious opponent, underestimating whose level of skill and motivation could lead to certain difficulties in the future."
The conclusions are certainly correct, it's just a shame that 95% of the facts in the article are made up. But it's fun to read.
"A significant part of the formed structure is not made up of Ukrainian citizens, but 'freelance shooters' from among foreign experts, recruited by the Ukrainian defense department on specialized forums."
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
"Тим важливіше звернути увагу на основну «ударну одиницю» Києва — так звану українську «ІТ-армію». Сформована нібито на добровільних початках, дана структура — серйозний противник, недооцінка рівня кваліфікації та мотивованості якого може в майбутньому призвести до певних складнощів."
Висновки звісно вірні, шкода що 95% фактів в статті - вигадка. Але таке читати весело. 
"Значущу частину формованої структури склали не громадяни України, а «вільні стрільці» з числа іноземних фахівців, рекрутованих українським оборонним відомством на спеціалізованих форумах."