Updated Info box for day 573
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Did you know that hacking has a century history? Back in 1903, Guglielmo Marconi was gearing up to showcase his wireless telegraph technology, transmitting a message from London to Cornwall. But a magician and prankster named Nevil Maskelyne hijacked the frequency and started spamming the word 'Rats,' followed by a cheeky poem about Marconi. 🎩✨
Maskelyne was the first hacker, proving that no technology is foolproof. So remember: we're part of a long, audacious tradition of keeping technology honest. 
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Знали ви, що хакерство має столітню історію? 1903 рік, Ґільєрмо Марконі готується продемонструвати свою бездротову телеграфну технологію передачі повідомлення з Лондона в Корнуолл. Але фокусник і жартівник Невіл Маскелін перехоплює частоту і починає спамити словом 'Rats', а потім відправляє нахабний вірш про Марконі. 🎩✨
Маскелін був першим хакером, який довів, що жодна технологія не є надійною. Тому пам'ятайте: ми є частиною довгої, нахабної традиції збереження чесності технологій. 👾