Added Day 574 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
"We understand perfectly well that the war will end, and we will win. But the war in cyberspace will never end," says Colonel Ivan Pavlenko, Ivan Pavlenko, Chief of the Main Directorate of Electronic Warfare and Cyber Security of the Ukrainian General Staff, in an interview with Ukrainska Pravda on September 12th. " Moreover, after their defeat on land, the "orcs" will focus even more on the cyber domain. So we should start preparing for that and increase our capacities right now."
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
«Ми прекрасно розуміємо, що війна закінчиться, ми переможемо. Але війна в кіберпросторі не закінчиться ніколи,» – говорить керівник Головного управління радіоелектронної боротьби та кібербезпеки Генерального штабу ЗСУ полковник Іван Павленко в інтерв’ю «Українській правді» 12 вересня. «Більше того, з програшем на землі «орки» ще більше полізуть у кіберсферу. Тому потрібно готуватися і нарощувати спроможності вже зараз.»