Added Day 588 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Ah, these Russian media and their fantastic tales! According to them, Ukraine has become a 'testing ground' for the West to experiment with large-scale cyber operations against poor Cinderella-Russia. And, of course, the USA is right there, being the main directors and founders of IT ARMY Ukraine. 🇺🇸😂
But here's the real 'gem': they claim we've targeted 6,000 Russian websites since February. Hilarious, we've even lost count, but it seems they've got it all under control. 😂😂😂
Have a great day, IT ARMY! 🛡️🌟

/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Ах, ці російські ЗМІ та їх фантастичні оповідання! За їхніми словами, Україна стала 'полігоном' для Заходу, щоб експериментувати з масштабними кіберопераціями проти попелюшки-росії. І, знову США тут як тут, є головними керівниками й засновниками IT ARMY України. 🇺🇸😂
Але ось головна 'перлина': вони стверджують, що ми атакували 6,000 російських сайтів з лютого. Смішно, навіть ми загубили рахунок, але у них, здається, все під контролем.😂😂😂
Вдалого дня, IT ARMY! 🛡️🌟