Updated Info box for day 590 | Оновлено Інформаційне поле для дня 590
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
We continue to disrupt the internet infrastructure in Crimea, which is actively used by the military occupiers. Currently, the provider Comfort XXI Century is in focus, whose services and internet have become hard to access for all clients. It's also important to disable IP, as enemy special services often use them as the main communication channel, as it was with Dom.Ru.
Meanwhile, our universal utility ADSS has received full support for Arch Linux. Joining our operations is becoming easier!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Продовжуємо вибивати інтернет інфраструктуру Криму, якою активно користуються військові-окупанти. Зараз під прицілом провайдер Комфорт ХХІ Век, сервіси й інтернет якого стали важкодоступними для всіх клієнтів. Виносити ІП також важливо, оскільки часто  ворожі спецслужби використовують їх як основний канал зв'язку, як це було з Дом.Ру.
Тим часом наша універсальна утиліта ADSS отримала повну підтримку Arch Linux. Приєднуватись до наших операцій стає все простіше!