Added Day 605 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Internet channels have become important components of military intelligence and coordination, not only for the Armed Forces of Ukraine but also for the enemy. However, lacking access to satellite internet technology, the enemy must rely on local infrastructure. Amid escalating military action in the Avdiivka area, we are helping our warriors.
Lots of IPs in the occupied territories, along with telecom operator Fenix, have faced significant issues with internet provision this week. Great work, friends!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Канали інтернет стали важливими компонентами військової розвідки та координації не тільки для ЗСУ, але і для ворога. Але ж ворог не маючи доступу до технології супутникового інтернету мусить використовувати локальну інфраструктуру. Тож на фоні загострення бойових дій в районі Авдіївки та не тільки, ми робимо все щоб допомогти нашим воїнам.
Ряд інтернет-провайдерів на окупованих територіях разом із телеком-оператором Фенікс, мали великі проблеми з наданням інтернету цього тижня. Чудова робота, друзі!