Added Day 610 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
🕰️ Cybercrime Flashback: In the early 2000s, the "ILOVEYOU" virus dominated global headlines. Originating from the Philippines, this worm spread via emails titled "ILOVEYOU" with an attached 'LOVE-LETTER-FOR-YOU.txt.vbs’ file. Upon opening, it rewrote files and self-propagated. Within ten days, it infected over 50 million computers, causing billions in damages.
Interested in creating content for the IT ARMY channel? Submit your post to armyuait@gmail.com
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
🕰️ Історія Кіберзлочинів: На початку 2000-х вірус "ILOVEYOU" став головним героєм на обкладинках світових новин. Походженням з Філіппін, цей комп'ютерний хробак передавався через електронну пошту з темою "ILOVEYOU" та доданим файлом 'LOVE-LETTER-FOR-YOU.txt.vbs’. Після відкриття він перезаписував файли та пересилав себе всім контактам. За всього десять днів він заразив понад 50 мільйонів комп'ютерів, завдавши збитків на мільярди.
Хочете створювати контент для каналу IT ARMY - відправте свій пост на armyuait@gmail.com