Added Day 614 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
🔥 Week #43: IT ARMY Operations Leaders 🏆
Unstoppable Micro_Monster leads again, this time in Distress and Mhddos! Welcome to the leaderboard, FreedomUA!
🛠 Mhddos: Micro_Monster 📊 427.4 TB
🛠 Distress: Micro_Monster 📊 844.2 TB
🛠 Db1000n: FreedomUA 📊 4.0 TB
🛠 X100: uashield21 📊 250.7 TB
Note: Different tools generate different amounts of traffic to achieve the desired effect, so they should not be compared across different tools.
Let's keep this spirit alive. On to new victories, IT ARMY! 🛡🤣
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
🔥 Тиждень #43: Лідери операцій IT ARMY 🏆
Нестримний Micro_Monster знову лідирує, цього разу в Distress та Mhddos! Вітаємо в лідерах FreedomUA!
🛠 Mhddos: Micro_Monster 📊 427.4 ТБ
🛠 Distress: Micro_Monster 📊 844.2 ТБ
🛠 Db1000n: FreedomUA 📊 4.0 ТБ
🛠 X100: uashield21 📊 250.7 ТБ
Зверніть увагу: Різні інструменти генерують різну кількість трафіку для досягнення потрібного ефекту, тож не варто їх порівнювати між різними інструментами.
Підтримуємо цей дух. До нових перемог, IT ARMY! 🛡🤣