Added Day 618 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Of course, it's commercial... Biden placed the order... only he hung up the phone on him. Said he'd settle with ATACMS. Asked where to send it. To us, or straight to Russia?
Congrats to the IT Army and our cyber intel unit for the most complex and prolonged attack on the Leonardo air traffic management system at the end of Sep and the airports earlier.
Russian banks were also the most attacked in the third quarter, and our contribution here was probably the most significant. Go forward, to cyber storm!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Звісно комерційна... Байден замовив... от тільки слухавку від нього поклав. Сказав Атакамсами розрахується. Питав куди відправляти. Нам, чи одразу на рассею?
А якщо серйозно, то вітаємо АйТі Військо й окремо команду кіберрозвідки з рекордною за складністю і тривалістю атаки на систему керування авіаперевезеннями Леонардо в кінці вересня та на інші аеропорти раніше. 
Російські банки також були найбільш атакованими в третьому кварталі, і наш внесок тут напевно був найвагомішим. Вперед, кіберштурм триває!