Updated Day 621 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
This time we made a move on the enemy's internet rear. INTERKOM, a provider which is one of the main suppliers of internet in central Moscow, where, as you know, there are many government, military, and other law enforcement offices. Let's see how their backup channels hold up and if anyone will operate through these channels...
Add an anonymous ID to your attacking tools. It's a great opportunity to preserve your contribution to the work of the IT ARMY. One day, after victory, we will be able to recognize everyone involved.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Цього разу пройшлися по інтернет тилам ворога. Провайдер ІНТЕРКОМ, який є одним з основних постачальників інтернету в центрі Москви, де як ви знаєте, багато офісів урядових, мілітарних, та й інших силових структур. Подивимось як їх резервні канали спрацюють, і чи не відпрацює хтось по цім каналам...
Додавайте анонімний ID до ваших атакуючих тулзів. Це чудова можливість зберегти свій внесок в роботу IT ARMY. Одного дня, після перемоги, ми зможемо відзначити всіх причетних.