Added Day 629 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
A critical "zero-day" vulnerability in SysAid identified as CVE-2023-47246 was exploited by cybercriminals linked to the CLOP ransomware. This security gap allowed the execution of arbitrary code. Microsoft detected the issue, and SysAid has already released a fix. Users affected are advised to immediately update to version 23.3.36 and follow SysAid's instructions to protect their systems.
Stay alert! Share problems and experiences with preventing cyber threats in our chat.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Критична вразливість "дня нуль" знайдена у SysAid, ідентифікована як CVE-2023-47246, була використана кіберзлочинцями, пов’язаними з вірусом-вимагачем CLOP. Ця безпекова шпарина дозволила виконання довільного коду. Microsoft виявив проблему, і SysAid вже випустив виправлення. Користувачам, що зазнали впливу, радимо негайно оновитися до версії 23.3.36 та слідувати вказівкам SysAid для захисту своїх систем. 
Будьте пильні! Діліться проблемами й досвідом з запобігання кіберзагроз у нашому чаті.