Added Day 639 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
The day has come! The phenomenon of the Ukrainian IT ARMY has become the subject of a fiction film. Starring: Robert De Niro as the State Department curator of the Army; Matt Damon as the main character, a hacker; and the hero's dog, a pug named Nabiy.
Although we weren't invited to consult on the cyber tricks or adapt the script to real events, it's incredibly gratifying that our example with you deserved to be the idea for a movie. Sincere thanks to the Ukrainians and everyone else who made this film!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Цей день настав! Феномен української IT ARMY став сюжетом художнього фільму. В ролях: Держдеп куратор Армії - Роберт Де Ніро; головний герой, хакер - Метт Демон; собака героя - мопс Набій.
І хоча нас не запросили консультувати кіно(кібер)трюки, чи адоптувати сценарій під реальні події - це неймовірно приємно що наш з вами приклад заслужив бути ідеєю для кіно. Щира подяка українцям і всім іншим, які зробили цю стрічку!