Added Day 642 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
We've disconnected a few more important internet providers in the occupied territories. Now it includes Nadezhda (Donetsk), Didan (Makiivka), and Inmart (Horlivka). An enemy without communication is assistance to our guys at the front. As a bonus, we slightly damaged the data in the DaData service, which even affected the operations of Russia's largest online retailer, Wildberries, and had an impact on some logistic chains.
So, shall we give the Muscovites a hellish week? Maybe this will be the last straw that breaks the camel's back.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Відключили ще декілька важливих інтернет-провайдерів на окупованих територіях. Тепер це Надежда (Донецьк), Дідан (Макіївка), Інмарт (Горлівка). Ворог без зв'язку - нашим хлопцям на фронті допомога. Бонусом трошки пошкодили дані в сервісі ДаДата, що відгукнулось навіть на роботі найбільшого онлайн ритейлера у РФ - Wildberries, та мало вплив на деякі логістичні ланцюжки.
Ну що влаштуємо москалям пекельний тиждень? Можливо це і буде та соломинка що верблюду хребет зламає.