Added Day 648 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
While experts continue to debate how to distinguish hackers from hacktivists, we offer a simple formula - just as a murderer differs from a soldier. IT ARMY is not only the world's first mass demonstration of hacktivism but also a consistent example of how important it is to create an ethical, and then legal, space for this phenomenon of modern cyber warfare.
It's gratifying that the SecurityWeek article cites the proactive stance of IT ARMY of Ukraine regarding openness to the regulation of cyber operations.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Хоча експерти досі сперечаються як відрізнити хакерів від хактивистів, ми пропонуємо просту формулу - так же само як вбивця відрізняється від солдата. IT ARMY це не тільки перший у світі масовий прояв хактивизму, але і послідовний приклад того наскільки важливо створити етичний, а потім і юридичний простір для цього явища сучасної кібервійни.
Приємно, що стаття SecurityWeek ставить у приклад проактивну позицію IT ARMY of Ukraine щодо  відкритості до врегулювання кібероперацій.