Added Day 661 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
lmatel and 2Kom, two major internet providers in the aggressor's capital, were disrupted this morning, impacting countless private, business, and government subscribers. This is all thanks to your efforts and our team's precise planning. While it's not as large as the KyivStar outage, it's a significant impact. The more devices you add, the more we can weaken the aggressor's infrastructure. Bring your friends into the IT ARMY! 🌐💻🛡️
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Алмател та 2Ком - два величезні інтернет-провайдери у столиці агресора перестали працювати сьогодні вранці. Це має бути десятки, якщо не сотні тисяч приватних, бізнес та урядових абонентів. Коли повернуться невідомо, але це все завдяки вашим зусиллям і ювелірної підготовки нашої команди розвідки. Це поки не масштаб збою КиївСтар, але хто знає яку операцію ми зможемо провести наступною?
Чим більше ви долучаєте девайсів, тим більше шкоди ми зможемо нанести інфраструктурі агресора. Приєднуйте до IT ARMY ваших друзів!