Added Day 669 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Bitrix has not yet fully recovered from our attack. Day four. This operation has become a certain milestone in the development of IT ARMY and the cherry on top in this challenging year. We congratulate everyone who has joined it!‍
‍Meanwhile, you can watch the DW report about the intricacies of modern cyber warfare. We are becoming witnesses and examples of how the perspective on the future war is being shaped. The future war that is currently going on in Ukraine.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Бітрікс досі повністю не відбився від нашої атаки. Четверта доба. Ця операція стала певною віхою у розвитку IT ARMY і вишнею на торті в цьому складному році. Вітаємо всіх і кожного хто до неї долучився!
А поки можете подивитися репортаж DW, про засади сучасної кібервійни. Ми стаємо свідками та прикладом того як формується погляд на війну майбутнього. Війну майбутнього яка зараз йде в Україні.