Added Day 675 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
1C, to fend off our attack, is creating a whitelist of client IPs manually. An effective method from thirty years ago. The only downside is that it's time-consuming, and until a client is added to the whitelist, they are disconnected from the service. And it can only be used by those with a static IP.
How many more of these Colossuses on clay feet will we find? But we need to constantly increase our capabilities. Friends, invite your friends to the IT ARMY!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
'Chameleon' - новий вірус для Android, що може обходити біометричний захист, становлячи загрозу для ваших особистих даних. Будьте обережні:
1С, щоб відбитися від нашої атаки створює вайтлист айпі клієнтів вручну. Дієвий спосіб тридцятирічної давнини. Тільки один мінус, це довго, а поки клієнт не внесений у вайтлист - він відключений від сервісу. І користуватися можна тільки маючи статичний айпі. 
Скільки ще таких Колоссів на глиняних ногах ми знайдемо? Але нам треба нарощувати потужності постійно. Друзі, звіть друзів у IT ARMY!