Added Day 777 - Targets and Days (TL) Eng+Ukr
Added Day 777 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Friends, our plan is now as follows: we conduct cyber operations, the Moskals are imprisoning each other, the RDK enters a country where half are imprisoned and the other half guard them. Or wait, hasn't this already happened?
Seriously though, there's currently an operation against another important enemy infrastructure, so check if all your irons are now whistling. We need more power!
And of course, get your friends and acquaintances to join the IT ARMY!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Друзі, план у нас тепер такий: ми проводимо кібероперації, москалі один одного саджають, РДК заходить в країну де половина сидить, а половина їх охороняє. Чи почекайте, це ж вже було?
Якщо серйозно, то зараз йде операція проти іншої важливої інфраструктури ворога, тож перевірте чи всі ваші праски зараз дудосять. Треба більше потужностей!
Ну і звісно приєднуйте до IT ARMY ваших друзів і знайомих!