Added Day 799 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Have you heard about mama's hackers? What about mama's cybersecurity experts? Here they are for your attention - advising their provider to hide their servers behind a VPN. Tip of the year. Ufanet couldn't have thought of this in seven days.
To the point, another provider, Digital Dialogue from the frontline Rostov region, was disconnected thanks to our efforts. Providers are the heart of the modern digital economy and often a crucial channel for military communications. So we continue our work!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Про мамкіних хакерів чули? А про мамкіних експертів з кібербезпеки? Ось вони до вашої уваги - радять своєму провайдеру увести свої сервери за VPN. Порада року. Уфанет до такого сім діб не міг додуматись.
По справі, ще один провайдер Цифровий Діалог з прифронтової Ростовської області був відключений завдяки нашим зусиллям. Провайдери це серце сучасної цифрової економіки та часто важливий канал військового зв'язку. Тому працюємо далі!