Added Day 802 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
The very Secretary of the enemy's Security Council was forced to respond to our post yesterday, threatening Western officials that by supporting the IT ARMY, they are opening Pandora's box which will eventually turn against its masters. It's nothing but a cry of Yaroslavna.
Furthermore, it turns out our MoDT has tracked us all down and prepared us in 'The Baltics', that's what they call the modern Baltic countries. Tomorrow, we will conduct a survey to find out who was prepared and who has ever visited the wonderful Lithuania, Latvia, and Estonia.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Сам секретар Радбеза ворога був вимушений реагувати на наш вчорашній пост, погрожуючи західним чиновникам, що вони підтримкою IT ARMY відкривають скриню Пандори та вона рано чи пізно повернеться проти своїх господарів. Плач Ярославни, та і тільки.
А ще виявляється наша Мінцифра всіх нас розшукала і підготувала у "Прибалтиці", це вони так сучасні країни Балтії кличуть. Будемо завтра опитування проводити, кого готували й хто взагалі коли-небудь гостював в чудових Литві, Латвії та Естонії.